backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Only needed for PostgreSQL - install separately if using PostgreSQL
python-dotenv==1.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
praw==7.7.1
tweepy==4.14.0
openai==1.3.5
chromadb==0.4.18
apscheduler==3.10.4
pydantic==2.5.0
pydantic-settings==2.1.0
sentence-transformers==2.2.2
torch==2.1.0
